LPS-50050 Update usage of ReflectionTestUtil.getFieldValue(Class, String)

+++ portal-service/test/unit/com/liferay/portal/kernel/io/OutputStreamWriterTest.java
@@ -209,7 +209,7 @@ public class OutputStreamWriterTest {
 	}
 
 	private int _getDefaultOutputBufferSize() {
-		return (Integer)ReflectionTestUtil.getFieldValue(
+		return ReflectionTestUtil.getFieldValue(
 			OutputStreamWriter.class, "_DEFAULT_OUTPUT_BUFFER_SIZE");
 	}
